custom encryption

+++ core/object/script_language.h
@@ -454,7 +454,7 @@ public:
 
 VARIANT_ENUM_CAST(ScriptLanguage::ScriptNameCasing);
 
-extern uint8_t script_encryption_key[32];
+extern uint16_t script_encryption_key[32];
 
 class PlaceHolderScriptInstance : public ScriptInstance {
 	Object *owner = nullptr;
